+++ MemControl/FCFS/FCFS.h
@@ -58,7 +58,7 @@ class FCFS : public MemoryController
 
     /* Stats */
     uint64_t measuredLatencies, measuredQueueLatencies;
-    float averageLatency, averageQueueLatency;
+    double averageLatency, averageQueueLatency;
     uint64_t mem_reads, mem_writes;
     uint64_t rb_hits;
     uint64_t rb_miss;
